♻️ Refactor: menu domain 위치 수정 (#96)
